Fix #17
This offers to fix https://github.com/affeldt-aist/coq2html/issues/17
by wiping out all end of line white spaces in block comments. The
drawback is that the end-of-line-two-spaces " " syntax of markdown to
force line breaks is no longer available in block comments. But at
least we get a nice and consistent output. The end-of-line-backslash
"\" syntax for line breaks remains available.