Setting max width to the content
Yet another small tweak to the CSS, setting max width to the content improves rendering of documentation. Right now the documentation expands to the full available space, and it doesn't look great with long lines rendered from markdown.
The parameter can be tweaked, 80ch enforces that the size should not expand over the 80 char limit, perhaps one can add a few characters of buffer, and even a parameter to customize this value.